Producer/Real/requirements.txt
kafka-python
requests
pytz